// packages/frontend/src/openapi/dnd-api/model/itemType.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * DnD
 * The DnD API description
 * OpenAPI spec version: 1.0
 */

export type ItemType = (typeof ItemType)[keyof typeof ItemType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ItemType = {
  Weapon: "Weapon",
  Spell: "Spell",
  ChestTrap: "ChestTrap",
  Potion: "Potion",
  Artifact: "Artifact",
} as const;